app/requirements.txt
psycopg2-binary==2.9.10
pypdf==5.4.0
python-dotenv==1.0.1